+++ gamemodes/terrortown/gamemode/shared/sh_init.lua
@@ -1,11 +1,11 @@
 ---
 -- This file contains all shared vars, tables and functions
 
-GM.Name = "TTT2 (Advanced Update)"
+GM.Name = "TTT2"
 GM.Author = "Bad King Urgrain, Alf21, saibotk, Mineotopia, LeBroomer, Histalek, ZenBre4ker"
 GM.Email = "ttt2@neoxult.de"
 GM.Website = "ttt.badking.net, docs.ttt2.neoxult.de"
-GM.Version = "0.12.0b"
+GM.Version = "0.12.1b"
 GM.Customized = true
 
 TTT2 = true -- identifier for TTT2. Just use "if TTT2 then ... end"