+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ## Unreleased
 
+### Fixed
+
+- Fixed the UI being unable to handle wrapping text with non-utf8 languages that do not use ASCII whitespaces (by @TimGoll & @saibotk)
+
 ## [v0.12.0b](https://github.com/TTT-2/TTT2/tree/v0.12.0b) (2023-12-11)
 
 ### Added
@@ -45,8 +49,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Added sql database access to panel elements
   - `DNumSliderTTT2`, `DCheckBoxLabelTTT2`, `DComboBoxTTT2`
 - Added dashing to propspec (by @TimGoll)
-- Added new functions to database module
-  -`database.SetDefaultValuesFromItem(accessName, itemName, item)`
+- Added new functions to database module -`database.SetDefaultValuesFromItem(accessName, itemName, item)`
 
 ### Changed
 
@@ -95,7 +98,6 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Optimized allocations by using global Vector / Angle when possible
 - Fixed spectated entity not being reset properly which can cause issues (by @TimGoll)
 - Optimized allocations by using global Vector / Angle when possible.
-- Fixed the UI being unable to handle languages that do not use ASCII whitespaces and therefore making text unreadable by placing it out of screen while also adding support for any unusual long words that would normally be cut off
 - Fixed the dynamic armor damage calculation being wrong when damage can only get partially reduced
 - Fixed propspec inputs behaving sometimes unexpectedly (by @TimGoll)
 - Fixed ComboBoxes not working with integer values (by @NickCloudAT)