+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Fixed ttt_game_text not working due to a refactor
 - Fixed dete call HUD being invisible
 - Fixed edgecase where undefined killer angle or pos were accessed
+- Fixed fallback ammo icon missing
 
 ## [v0.12.0b](https://github.com/TTT-2/TTT2/tree/v0.12.0b) (2023-12-11)