+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Fixed dete call HUD being invisible
 - Fixed edgecase where undefined killer angle or pos were accessed
 - Fixed fallback ammo icon missing
+- Fixed a null entity error in the miniscoreboard
 
 ## [v0.12.0b](https://github.com/TTT-2/TTT2/tree/v0.12.0b) (2023-12-11)